Implementação do pattern View Model com AutoMapper. Criação novo json converter customizado para decimal com 2 casas. Configuração do Mvc

+++ src/Services/CalculoJuros/MMM.Teste.CalculoJuros.Application/Interfaces/ITaxaJurosService.cs
@@ -4,6 +4,6 @@ namespace MMM.Teste.CalculoJuros.Application.Services
 {
     public interface ITaxaJurosService
     {
-        Task<double?> GetTaxaJuros();
+        Task<decimal?> GetTaxaJuros();
     }
 }
